refactor test class to split into two partial classes.
